refactor(tools): import Tool/TOOLS from src/tools.ts directly
Address Copilot review on PR #70: args.ts both re-exported and
imported TOOLS/Tool from ./tools.ts, which works but reads as
clunky redundancy. Drop the re-export and update the four
consumers (cli.ts, prompt.ts, telemetry.ts, workspace.ts) to
import the type from tools.ts directly. Single source of truth;
args.ts no longer pretends to own the symbol.
